sponsor page is now a functional component:
